app[REFACT] moved signUpFetch in VM
