+++ packages/app/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./useFetchAuth";
 export * from "./useTextInput";
 export * as signInScreenVM from "./signInScreenVM";
+export * as signUpScreenVM from "./signUpScreenVM";
